+++ src/androidTest/kotlin/us/frollo/frollosdk/aggregation/AggregationTest.kt
@@ -448,31 +448,39 @@ class AggregationTest : BaseAndroidTest() {
     }
 
     @Test
-    fun testForceRefreshProviderAccounts() {
+    fun testSyncProviderAccounts() {
         initSetup()
 
-        val body = readStringFromJson(app, R.raw.provider_accounts_valid)
+        val refreshBody = readStringFromJson(app, R.raw.provider_accounts_valid)
+        val syncBody = readStringFromJson(app, R.raw.provider_accounts_valid_sync)
         mockServer.setDispatcher(object : Dispatcher() {
             override fun dispatch(request: RecordedRequest?): MockResponse {
                 if (request?.trimmedPath?.contains(AggregationAPI.URL_PROVIDER_ACCOUNTS) == true) {
                     return MockResponse()
                             .setResponseCode(200)
-                            .setBody(body)
+                            .setBody(if (request.method == "PUT") syncBody else refreshBody)
                 }
                 return MockResponse().setResponseCode(404)
             }
         })
 
-        val array = longArrayOf(623L, 624L, 625L, 864L)
-        aggregation.forceRefreshProviderAccounts(array) { result ->
-            assertEquals(Result.Status.SUCCESS, result.status)
-            assertNull(result.error)
+        aggregation.refreshProviderAccounts {
+            assertEquals(Result.Status.SUCCESS, it.status)
+            assertNull(it.error)
 
-            val testObserver = aggregation.fetchProviderAccounts().test()
-            testObserver.awaitValue()
-            val models = testObserver.value().data
-            assertNotNull(models)
-            assertEquals(4, models?.size)
+            val array = longArrayOf(623L, 624L)
+            aggregation.syncProviderAccounts(array) { result ->
+                assertEquals(Result.Status.SUCCESS, result.status)
+                assertNull(result.error)
+
+                val testObserver = aggregation.fetchProviderAccounts().test()
+                testObserver.awaitValue()
+                val models = testObserver.value().data
+                assertNotNull(models)
+                assertEquals(4, models?.size)
+                assertEquals(AccountRefreshStatus.FAILED, models?.get(0)?.refreshStatus?.status)
+                assertEquals(AccountRefreshStatus.UPDATING, models?.get(1)?.refreshStatus?.status)
+            }
         }
 
         wait(3)